chore(cas): Cleanup logger to prepare it to be its own package
